Update guest to load shared utils from app (#2617)
* Update guest to load shared utils from app

* Remove unused import

* Update imports

* Update rollup.config.js

* Update rollup.config.js

* Update ui/app/src/utils/contentType.ts
