+++ ui/app/src/utils/array.ts
@@ -16,21 +16,22 @@
 
 import LookupTable from '../models/LookupTable';
 
-export function forEach<T = any>(
+export function forEach<T = any, R = undefined>(
   array: T[],
-  fn: (item: T, index: number, array: T[]) => any,
-  emptyReturnValue: any = undefined
-): any {
+  fn: (item: T, index: number, array: T[]) => R | 'continue' | 'break' | undefined,
+  emptyReturnValue?: R
+): R {
   if (emptyReturnValue != null && array?.length === 0) {
     return emptyReturnValue;
   }
   for (let i = 0, l = array.length; i < l; i++) {
     const result = fn(array[i], i, array);
-    if (result === 'continue') {
-    } else if (result === 'break') {
-      break;
-    } else if (result !== undefined) {
-      return result;
+    if (result !== 'continue') {
+      if (result === 'break') {
+        break;
+      } else if (result !== undefined) {
+        return result;
+      }
     }
   }
   return emptyReturnValue;
@@ -63,4 +64,17 @@ export function createPresenceTable<T extends any = boolean, K extends any = ''>
       : (value) => (table[itemKeyGenerator(value)] = valueOrGenerator);
   list.forEach(callback);
   return table;
+}
+
+/**
+ * Receives two different arrays of string, it will combine it selecting one string of the first array
+ * and then other string of the second array, until there are no more strings
+ * @param a
+ * @param b
+ */
+export function mergeArraysAlternatively(a: string[], b: string[]): string[] {
+  return (a.length > b.length ? a : b).reduce(
+    (acc, cur, i) => (a[i] && b[i] ? [...acc, a[i], b[i]] : [...acc, cur]),
+    []
+  );
 }
