+++ ui/guest/src/index.tsx
@@ -19,7 +19,7 @@ import ReactDOM from 'react-dom';
 import Guest, { GuestProps } from './react/Guest';
 import GuestProxy from './react/GuestProxy';
 import ContentInstance from '@craftercms/studio-ui/models/ContentInstance';
-import { nnou } from './utils/object';
+import { nnou } from '@craftercms/studio-ui/utils/object';
 import * as elementRegistry from './classes/ElementRegistry';
 import * as iceRegistry from './classes/ICERegistry';
 import * as contentController from './classes/ContentController';
