+++ ui/guest/src/react/DropMarker.tsx
@@ -18,8 +18,8 @@ import React, { useEffect, useRef, useState } from 'react';
 import { foo, HORIZONTAL, VERTICAL, X_AXIS, Y_AXIS } from '../utils/util';
 import { Coordinates } from '../models/Positioning';
 import { DropZone, ElementRecord } from '../models/InContextEditing';
-import { notNullOrUndefined } from '../utils/object';
-import { forEach } from '../utils/array';
+import { notNullOrUndefined } from '@craftercms/studio-ui/utils/object';
+import { forEach } from '@craftercms/studio-ui/utils/array';
 import { findClosestRect, getDropMarkerPosition, getInRectStats, splitRect } from '../utils/dom';
 import { FullSxRecord, PartialSxRecord } from '@craftercms/studio-ui/models/CustomRecord';
 import { Box } from '@mui/material';
