+++ ui/guest/src/react/GuestProxy.tsx
@@ -30,7 +30,7 @@ import {
 } from '../classes/ContentController';
 import { zip } from 'rxjs';
 import { filter, take } from 'rxjs/operators';
-import * as ContentType from '../utils/contentType';
+import * as ContentType from '@craftercms/studio-ui/utils/contentType';
 import { message$ } from '../utils/communicator';
 import { Operation } from '../models/Operations';
 import {
@@ -42,8 +42,8 @@ import {
   updateFieldValueOperation
 } from '@craftercms/studio-ui/state/actions/preview';
 import { GuestState } from '../store/models/GuestStore';
-import { notNullOrUndefined } from '../utils/object';
-import { forEach } from '../utils/array';
+import { notNullOrUndefined } from '@craftercms/studio-ui/utils/object';
+import { forEach } from '@craftercms/studio-ui/utils/array';
 import { popPiece, removeLastPiece } from '@craftercms/studio-ui/utils/string';
 import { addAnimation } from '../utils/dom';
