+++ ui/guest/src/store/epics/root.ts
@@ -50,7 +50,7 @@ import {
 } from '@craftercms/studio-ui/state/actions/preview';
 import { MouseEventActionObservable } from '../models/Actions';
 import { GuestState } from '../models/GuestStore';
-import { isNullOrUndefined, notNullOrUndefined, reversePluckProps } from '../../utils/object';
+import { nullOrUndefined, notNullOrUndefined, reversePluckProps } from '@craftercms/studio-ui/utils/object';
 import { ElementRecord, ICEProps } from '../../models/InContextEditing';
 import * as ElementRegistry from '../../classes/ElementRegistry';
 import { get } from '../../classes/ElementRegistry';
@@ -91,7 +91,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
           payload: { event, record }
         } = action;
         const iceId = state.draggable?.[record.id];
-        if (isNullOrUndefined(iceId)) {
+        if (nullOrUndefined(iceId)) {
           // When the drag starts on a child element of the item, it passes through here.
           console.error('No ice id found for this drag instance.', record, state.draggable);
         } else if (not(iceId)) {
@@ -513,7 +513,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
       withLatestFrom(state$),
       switchMap(([, state]) => {
         const contentType = state.dragContext.contentType;
-        if (isNullOrUndefined(contentType.id)) {
+        if (nullOrUndefined(contentType.id)) {
           console.error('No contentTypeId found for this drag instance.');
         } else {
           if (state.dragContext.dropZones.length === 0) {
@@ -537,7 +537,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
       ofType(componentInstanceDragStarted.type),
       withLatestFrom(state$),
       switchMap(([, state]) => {
-        if (isNullOrUndefined(state.dragContext.instance.craftercms.contentTypeId)) {
+        if (nullOrUndefined(state.dragContext.instance.craftercms.contentTypeId)) {
           console.error('No contentTypeId found for this drag instance.');
         } else {
           if (state.dragContext.dropZones.length === 0) {
@@ -562,7 +562,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
       ofType(assetDragStarted.type),
       withLatestFrom(state$),
       switchMap(([, state]) => {
-        if (isNullOrUndefined(state.dragContext.dragged.path)) {
+        if (nullOrUndefined(state.dragContext.dragged.path)) {
           console.error('No path found for this drag asset.');
         } else {
           return initializeDragSubjects(state$);
@@ -578,7 +578,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
       ofType(desktopAssetDragStarted.type),
       withLatestFrom(state$),
       switchMap(([, state]) => {
-        if (isNullOrUndefined(state.dragContext.dragged)) {
+        if (nullOrUndefined(state.dragContext.dragged)) {
           console.error('No file found for this drag asset.');
         } else {
           return initializeDragSubjects(state$);
